Fixes for newer verison of Python
There is no longer a xlsxwriter.save(), only a xlsxwriter.close().

Seems that a new version of pandas doesn't allow the index to be
remerged into the multi-index column when that same multi-index
column name exists already, so we can just drop the index without
putting it back into a column.